Rename 'fluent' package to 'base'
