+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/GndFirestore.java
@@ -28,9 +28,9 @@ import com.google.android.gnd.model.feature.FeatureMutation;
 import com.google.android.gnd.model.observation.Observation;
 import com.google.android.gnd.model.observation.ObservationMutation;
 import com.google.android.gnd.persistence.remote.RemoteDataEvent;
-import com.google.android.gnd.persistence.remote.firestore.fluent.FluentCollectionReference;
-import com.google.android.gnd.persistence.remote.firestore.fluent.FluentDocumentReference;
-import com.google.android.gnd.persistence.remote.firestore.fluent.FluentFirestore;
+import com.google.android.gnd.persistence.remote.firestore.base.FluentCollectionReference;
+import com.google.android.gnd.persistence.remote.firestore.base.FluentDocumentReference;
+import com.google.android.gnd.persistence.remote.firestore.base.FluentFirestore;
 import com.google.common.collect.ImmutableList;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentChange;
