+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/base/FluentDocumentReference.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.persistence.remote.firestore.fluent;
+package com.google.android.gnd.persistence.remote.firestore.base;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.firebase.firestore.DocumentReference;
